requirements.txt
azure-eventhub==5.11.*
pyserial==3.*
